move scrollTo to dom module
