+++ src/js/lib/dom.js
@@ -10,4 +10,16 @@ function setStyles(el, styles) {
     Object.keys(styles).forEach(style => el.style[style] = styles[style]);
 }
 
-export {setAttributes, setData, setStyles};
+function scrollTo(element, to, duration) {
+    if (duration <= 0) return;
+    const difference = to - element.scrollTop;
+    const perTick = difference / duration * 10;
+
+    setTimeout(() => {
+        element.scrollTop = element.scrollTop + perTick;
+        if (element.scrollTop === to) return;
+        scrollTo(element, to, duration - 10);
+    }, 10);
+}
+
+export {setAttributes, setData, setStyles, scrollTo};
